Split the management API into two parts, one that is generic and has no native API dependency and another that does have a native API dependency.
